+++ crates/rslint_cli/src/main.rs
@@ -1,5 +1,14 @@
 use rslint_cli::ExplanationRunner;
-use structopt::StructOpt;
+use structopt::{clap::arg_enum, StructOpt};
+
+const DEV_FLAGS_HELP: &str = "
+Developer flags that are used by RSLint developers to debug RSLint.
+
+    -Z help     -- Shows this message
+    -Z tokenize -- Tokenizes the input files and dumps the tokens
+    -Z dumpast  -- Parses the input files and prints the parsed AST
+
+Run with 'rslint -Z <FLAG> <FILES>'.";
 
 #[derive(Debug, StructOpt)]
 #[structopt(
@@ -12,7 +21,7 @@ pub(crate) struct Options {
     verbose: bool,
     /// A glob pattern to lint.
     #[structopt(default_value = "./")]
-    files: String,
+    files: Vec<String>,
     #[structopt(subcommand)]
     cmd: Option<SubCommand>,
     /// Automatically attempt to fix any issues which can be fixed
@@ -21,9 +30,24 @@ pub(crate) struct Options {
     /// Attempt to run autofixes even if the code contains syntax errors (may produce weird fixes or more errors)
     #[structopt(short = "D", long)]
     dirty: bool,
+    /// Disables the global config that is located in your global config directory.
+    #[structopt(long)]
+    no_global_config: bool,
     /// The error formatter to use, either "short" or "long" (default)
     #[structopt(short = "F", long)]
     formatter: Option<String>,
+    /// Developer only flags. See `-Z help` for more information.
+    #[structopt(name = "FLAG", short = "Z")]
+    dev_flag: Option<DevFlag>,
+}
+
+arg_enum! {
+    #[derive(Debug, PartialEq, Eq)]
+    enum DevFlag {
+        Help,
+        Tokenize,
+        DumpAst,
+    }
 }
 
 #[derive(Debug, StructOpt, PartialEq, Eq)]
@@ -33,6 +57,8 @@ pub(crate) enum SubCommand {
     /// Show all of the available rules
     // TODO: show only rules of particular groups
     Rules,
+    /// Try to infer the options of some rules from various files and print the results
+    Infer { files: Vec<String> },
 }
 
 fn main() {
@@ -41,11 +67,21 @@ fn main() {
 
     let opt = Options::from_args();
 
-    if let Some(SubCommand::Explain { rules }) = opt.cmd {
-        ExplanationRunner::new(rules).print();
-    } else if opt.cmd == Some(SubCommand::Rules) {
-        rslint_cli::show_all_rules();
-    } else {
-        rslint_cli::run(opt.files, opt.verbose, opt.fix, opt.dirty, opt.formatter);
+    match (opt.dev_flag, opt.cmd) {
+        (Some(DevFlag::Help), _) => println!("{}", DEV_FLAGS_HELP),
+        (Some(DevFlag::Tokenize), _) => rslint_cli::tokenize(opt.files),
+        (Some(DevFlag::DumpAst), _) => rslint_cli::dump_ast(opt.files),
+
+        (_, Some(SubCommand::Explain { rules })) => ExplanationRunner::new(rules).print(),
+        (_, Some(SubCommand::Rules)) => rslint_cli::show_all_rules(),
+        (_, Some(SubCommand::Infer { files })) => rslint_cli::infer(files),
+        (_, None) => rslint_cli::run(
+            opt.files,
+            opt.verbose,
+            opt.fix,
+            opt.dirty,
+            opt.formatter,
+            opt.no_global_config,
+        ),
     }
 }