+++ crates/rslint_cli/CHANGELOG.md
@@ -6,11 +6,30 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Changed the binary's name from `rslint_cli` to just `rslint`
+- Changed the exit codes produced by the CLI:
+  `-1`: Linting was unsuccessful because of an internal error
+  `0`: Linting was successful and there are no errors (but there may be warnings)
+  `1`: Linting was successful and there is at least one error
+  `2`: Linting could not be done because of a config or CLI error (e.g. invalid glob pattern)
+- Changed CLI to accept multiple glob patterns
+
 ### Added
 
 - Added ways of configuring the formatter used through CLI and config
 - Made all fields of config public
 - Added the `rules` subcommand to show all available rules
+- Added `JsFile::parse`
+- Changed the parameters of `FileWalker::from_glob` from `Paths` to a generic `IntoIterator`
+- Added the `infer` subcommand
+- Added the `infer` function
+- `-Z` developer flags (`dumpast`, `tokenize`, `help`)
+
+### Changed
+
+- Move config handling to `rslint_config`
 
 ## [0.2.1] - 2020-10-21