+++ crates/rslint_core/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added `util::StyleExt` for stylistic linting
 - Added `Fixer::delete_multiple`
+- Added more functions to `util::StyleExt`
+- Added `RuleCtx::dummy_ctx`
+- Added `Inferable` to define rules which can have their options inferred from nodes
+- Added benchmarks for linting a file
 
 ### Changed